requirements.txt
fastapi
uvicorn
pandas
openpyxl
pydantic
requests
starlette.responses